+++ PokemonGo-UWP/Utils/Signature/DeviceInfosIOS.cs
@@ -185,7 +185,7 @@ namespace PokemonGo_UWP.Utils
             public double AccelRawZ => _accelerometer?.GetCurrentReading()?.AccelerationZ ?? _random.NextGaussian(0.0, 0.3);
 
 
-            public ulong AccelerometerAxes => 3;
+            public int AccelerometerAxes => 3;
 
 
         }
@@ -231,7 +231,7 @@ namespace PokemonGo_UWP.Utils
                 // TODO: why 1? need more infos.
                 loc.LocationType = 1;
 
-                loc.TimeSnapshot = DeviceInfos.RelativeTimeFromStart;
+                loc.TimeSnapshot = (ulong)DeviceInfos.RelativeTimeFromStart;
 
                 loc.HorizontalAccuracy = (float?)LocationServiceHelper.Instance.Geoposition.Coordinate?.Accuracy ?? (float)_random.NextDouble(5.0, 50.0);
 
@@ -256,7 +256,7 @@ namespace PokemonGo_UWP.Utils
 
             public ulong LocationType { get; private set; }
 
-            public long TimeSnapshot { get; private set; }
+            public ulong TimeSnapshot { get; private set; }
             public float HorizontalAccuracy { get; private set; }
             public float VerticalAccuracy { get; private set; }
             public float Course { get; private set; }
