+++ PokemonGoAPI/Client.cs
@@ -40,7 +40,7 @@ namespace PokemonGo.RocketAPI
 
         public double CurrentLatitude { get; internal set; }
         public double CurrentLongitude { get; internal set; }
-        public double CurrentAccuracy { get; internal set; }
+        public int CurrentAccuracy { get; internal set; }
 
         public AuthType AuthType => Settings.AuthType;
         internal string ApiUrl { get; set; }
